Add files via upload
Version 0.9a

Fixes: 
Fixed loud initialization using hold boolean on the linear interpolating delay until the Dattorro has reached equilibrium
Implemented knobs
Implemented CVs
Fixed some heavy static as result of unoptimized code, unstable knobs, and sample rate choices.
Longer predelay buffer

To-do:
Implement low cut and high cut filters on the tank and on the inputs
Implement freeze function

+++ dsp/delays/AllpassFilter.hpp
@@ -1,9 +1,3 @@
-/**
- * @file AllpassFilter.hpp
- * @author Dale Johnson
- * @date ...uhhh
- */
-
 #pragma once
 #include "InterpDelay.hpp"
 
@@ -14,7 +8,7 @@ public:
         gain = 0.;
     }
 
-    AllpassFilter(long maxDelay, long initDelay = 0, float gain = 0.) {
+    AllpassFilter(int maxDelay, int initDelay = 0, float gain = 0.) {
         clear();
         delay = InterpDelay(maxDelay, initDelay);
         this->gain = gain;
@@ -49,53 +43,3 @@ private:
     float _inSum;
     float _outSum;
 };
-
-// template<class T>
-// class NestedAllPassType1 {
-// public:
-//     NestedAllPassType1() {
-//         gain1 = 0;
-//         gain2 = 0;
-//         decay1 = 0;
-//         decay2 = 0;
-//         clear();
-//     }
-
-//     NestedAllPassType1(long maxDelay, long delayTime1, long delayTime2) {
-//         delay1 = InterpDelay(maxDelay, delayTime1);
-//         delay2 = InterpDelay(maxDelay, delayTime2);
-//         gain1 = 0;
-//         gain2 = 0;
-//         decay1 = 0;
-//         decay2 = 0;
-//         clear();
-//     }
-
-//     float inline process() {
-//         _inSum1 = input + delay1.output * gain1;
-//         _inSum2 = _inSum1 + delay2.output * gain2;
-//         delay2.input = _inSum2;
-//         delay1.input = delay2.output * decay2 + _inSum2 * -gain2;
-//         output = delay1.output * decay1 + _inSum1 * -gain1;
-//         delay1.process();
-//         delay2.process();
-//         return output;
-//     }
-
-//     void clear() {
-//         input = 0;
-//         output = 0;
-//         _inSum1 = 0;
-//         _inSum2 = 0;
-//         delay1.clear();
-//         delay2.clear();
-//     }
-
-//     float input;
-//     float gain1, gain2;;
-//     float output;
-//     float decay1, decay2;
-//     InterpDelay delay1, delay2;
-// private:
-//     float _inSum1, _inSum2;
-// };
